feat!: convert whole project to a Flask project
Manually grabbing the events.json file is quite
tiring, especially when knowing the fact that
Simaster's calendar changes quite often.

This commit converted the whole project.
This project is now a web-based app which has an
API to grab a specific Simaster account's
calendar and return it as an iCalendar file.

Signed-off-by: Faiz Jazadi <me@lcat.dev>